GW-1577: Added approved field to get payment response (#45)
