+++ src/CheckoutSdk/Payments/GetPaymentResponse.cs
@@ -14,84 +14,89 @@ namespace Checkout.Payments
         /// Gets the unique identifier of the payment.
         /// </summary>
         public string Id { get; set; }
-        
+
         /// <summary>
         /// Gets the date/time the payment was requested.
         /// </summary>
         public DateTime RequestedOn { get; set; }
-        
+
         /// <summary>
         /// Gets the source of the payment.
         /// </summary>
         [JsonConverter(typeof(ResponseSourceConverter))]
         public IResponseSource Source { get; set; }
-        
+
         /// <summary>
         /// Gets the original payment amount.
         /// </summary>
         public int? Amount { get; set; }
-        
+
         /// <summary>
         /// Gets the three-letter ISO currency code of the payment.
         /// </summary>
         public string Currency { get; set; }
-        
+
         /// <summary>
         /// Gets the payment type.
         /// </summary>
         public PaymentType PaymentType { get; set; }
-        
+
         /// <summary>
         /// Gets your reference for the payment.
         /// </summary>
         public string Reference { get; set; }
-        
+
         /// <summary>
         /// Gets your description of the payment.
         /// </summary>
         public string Description { get; set; }
-        
+
+        /// <summary>
+        /// Gets whether the payment was successfully authorized or captured.
+        /// </summary>
+        public bool Approved { get; set; }
+
         /// <summary>
         /// Gets the status of the payment.
         /// </summary>
         public PaymentStatus Status { get; set; }
-        
+
         /// <summary>
         /// Gets 3D-Secure information relating to the payment.
         /// </summary>
         [JsonProperty(PropertyName = "3ds")]
         public ThreeDSEnrollment ThreeDS { get; set; }
-        
+
         /// <summary>
         /// Gets the payment's risk response.
         /// </summary>
         public RiskAssessment Risk { get; set; }
-        
+
         /// <summary>
         /// Gest the customer to which this payment is linked.
         /// </summary>
         public CustomerResponse Customer { get; set; }
-        
+
         /// <summary>
         /// Gets the billing descriptor displayed on the account owner's statement.
         /// </summary>
         public BillingDescriptor BillingDescriptor { get; set; }
-        
+
         /// <summary>
         /// Gets the payment's shipping details.
         /// </summary>
         public ShippingDetails Shipping { get; set; }
-        
+
         /// <summary>
         /// Gets the customer IP address used to make the payment.
         /// </summary>
         public string PaymentIp { get; set; }
-        
+
         /// <summary>
         /// Gets the payment recipient details (Required by VISA and MasterCard for domestic UK transactions processed by Financial Institutions).
         /// </summary>
         public PaymentRecipient Recipient { get; set; }
-        
+
         /// <summary>
         /// Gets the metadata you attached to the original payment request.
         /// </summary>
@@ -107,7 +112,7 @@ namespace Checkout.Payments
         /// </summary>
         /// <returns>True if a redirect is required, otherwise False.</returns>
         public bool RequiresRedirect() => HasLink("redirect");
-        
+
         /// <summary>
         /// Gets the redirect link.
         /// </summary>
